style(msvc): fix C4244/C4267 in every remaining 4-site file (item 2)
All static_cast, documenting existing intentional narrowing -- no
logic changes. sm_tests unchanged at 5966/226.

- StepMania.cpp (4): a ceil() result into a window-width int; a
  MEMORYSTATUS::dwTotalPhys division into an int; an
  srand(time(nullptr)) seed narrowing.
- SongUtil.cpp (4): two SecondsToMMSS(int) calls in a length-sort
  label; an RString::Left fed a size_t-minus arg; a reverse-loop
  bound.
- ScreenUnlockStatus.cpp (4): two size_t unlock counts into unsigned;
  two reverse-loop bounds.
- ScreenServiceAction.cpp (4): two near-duplicate edit-clearing
  helpers each with a size_t file count and a count_if result into
  int.
- ScreenJukebox.cpp (4): three RandomInt(size_t) calls; a
  reverse-loop bound.
- ScoreKeeperNormal.cpp (4): a size_t song-count into int; two
  lua_tointeger()-into-int sites in a toasty-trigger callback.
- RageTimer.cpp (4): two Difference() (returns double) results
  narrowed to the class's float-returning API; a std::floor(float)
  result into int64_t seconds.
- RageSoundReader_Preload.cpp (4): four m_Buffer->size() / framesize
  sites (a repeated frame-count computation) into int.
- RageSoundReader_MP3.cpp (4): three MAD-stream pointer-diffs into
  int; an id3_tag_query buffer-length arg into id3_length_t.
- RageMath.cpp (4): a ternary using double literals where the function
  returns float (fixed with f-suffixed literals instead of a cast);
  three double-typed intermediate results in a triangle-wave helper
  cast at each return.
- RageDisplay_D3D.cpp (4): a palette-index map lookup's size_t value
  into int; two std::max<unsigned int> calls fed size_t vertex/
  triangle totals (same pattern as the earlier OGL fix).

+++ src/RageSoundReader_Preload.cpp
@@ -45,7 +45,7 @@ RageSoundReader_Preload::RageSoundReader_Preload():
 
 int RageSoundReader_Preload::GetTotalFrames() const
 {
-	return m_Buffer->size() / framesize;
+	return static_cast<int>(m_Buffer->size() / framesize);
 }
 
 bool RageSoundReader_Preload::Open( RageSoundReader *pSource )
@@ -130,7 +130,7 @@ int RageSoundReader_Preload::SetPosition( int iFrame )
 
 	if( m_iPosition >= int(m_Buffer->size() / framesize) )
 	{
-		m_iPosition = m_Buffer->size() / framesize;
+		m_iPosition = static_cast<int>(m_Buffer->size() / framesize);
 		return 0;
 	}
 
@@ -144,7 +144,7 @@ int RageSoundReader_Preload::GetNextSourceFrame() const
 
 int RageSoundReader_Preload::Read( float *pBuffer, int iFrames )
 {
-	const int iSizeFrames = m_Buffer->size() / framesize;
+	const int iSizeFrames = static_cast<int>(m_Buffer->size() / framesize);
 	const int iFramesAvail = iSizeFrames - m_iPosition;
 
 	iFrames = std::min( iFrames, iFramesAvail );
